+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [0.5.15] - 2026-03-03
+
+### Summary
+
+refactor(code2llm): code analysis engine
+
+### Other
+
+- update code2llm/core/analyzer.py
+- update code2llm/core/core/refactoring.py
+
+
 ## [0.5.13] - 2026-03-02
 
 ### Summary